Various changes
Login string may not be an email
Made the "Recover password" functionality configurable: you can enable or disable it.
Added the 'admin_key' and 'password_recovery' config keys to the skylable.ini config file.
Updated the installer to let the user set the password recovery functionality
Updated the login code: on new user check if the login is an email and save it into the db.